+++ buildSrc/src/main/kotlin/dependency/ui/AndroidCore.kt
@@ -6,5 +6,5 @@ object AndroidCore : Dependency {
 
     override val group: String = "androidx.core"
     override val artifact: String = "core"
-    override val version: String = "1.1.0"
+    override val version: String = "1.3.0"
 }