+++ buildSrc/src/main/kotlin/dependency/ui/Paging.kt
@@ -6,5 +6,5 @@ object Paging : Dependency {
 
     override val group = "androidx.paging"
     override val artifact = "paging-runtime"
-    override val version = "2.1.0"
+    override val version = "2.1.2"
 }