+++ helper/src/commonMain/kotlin/com/algolia/instantsearch/helper/filter/range/FilterRangeConnection.kt
@@ -6,8 +6,8 @@ import com.algolia.instantsearch.core.number.range.connectView
 import com.algolia.instantsearch.helper.filter.state.FilterGroupID
 import com.algolia.instantsearch.helper.filter.state.FilterOperator
 import com.algolia.instantsearch.helper.filter.state.FilterState
+import com.algolia.instantsearch.helper.searcher.SearcherSingleIndex
 import com.algolia.search.model.Attribute
-
 
 public fun <T> FilterRangeViewModel<T>.connectFilterState(
     filterState: FilterState,
@@ -21,4 +21,60 @@ public fun <T> FilterRangeConnector<T>.connectView(
     view: NumberRangeView<T>
 ): Connection where T : Number, T : Comparable<T> {
     return viewModel.connectView(view)
+}
+
+/**
+ * Create a connection between a searcher and the filter range components for dynamic behavior.
+ *
+ * @param searcher searcher for facets and results extraction operations.
+ * @param attribute attribute to dynamically use its facet stats results.
+ */
+public inline fun <reified T> FilterRangeConnector<T>.connectSearcher(
+    searcher: SearcherSingleIndex,
+    attribute: Attribute
+): Connection where T : Number, T : Comparable<T> {
+    return viewModel.connectSearcher(searcher, attribute)
+}
+
+/**
+ * Create a connection between a searcher and the filter range components for dynamic behavior.
+ *
+ * @param searcher searcher for facets and results extraction operations.
+ * @param attribute attribute to dynamically use its facet stats results.
+ */
+public inline fun <reified T> FilterRangeViewModel<T>.connectSearcher(
+    searcher: SearcherSingleIndex,
+    attribute: Attribute
+): Connection where T : Number, T : Comparable<T> {
+    return FilterRangeConnectionSearcher(this, searcher, attribute, T::class)
+}
+
+/**
+ * Create a connection between a searcher and the filter range components for dynamic behavior.
+ *
+ * @param searcher searcher for facets and results extraction operations.
+ * @param attribute attribute to dynamically use its facet stats results.
+ * @param mapper explicit mapper to transform facets stats min/max results to the view model's bounds.
+ */
+public inline fun <reified T> FilterRangeConnector<T>.connectSearcher(
+    searcher: SearcherSingleIndex,
+    attribute: Attribute,
+    noinline mapper: (Number) -> T
+): Connection where T : Number, T : Comparable<T> {
+    return viewModel.connectSearcher(searcher, attribute, mapper)
+}
+
+/**
+ * Create a connection between a searcher and the filter range components for dynamic behavior.
+ *
+ * @param searcher searcher for facets and results extraction operations.
+ * @param attribute attribute to dynamically use its facet stats results.
+ * @param mapper explicit mapper to transform facets stats min/max results to the view model's bounds.
+ */
+public inline fun <reified T> FilterRangeViewModel<T>.connectSearcher(
+    searcher: SearcherSingleIndex,
+    attribute: Attribute,
+    noinline mapper: (Number) -> T
+): Connection where T : Number, T : Comparable<T> {
+    return FilterRangeConnectionSearcher(this, searcher, attribute, mapper)
 }