+++ helper/src/commonMain/kotlin/com/algolia/instantsearch/helper/hierarchical/HierarchicalConnectionSearcher.kt
@@ -6,7 +6,6 @@ import com.algolia.instantsearch.helper.searcher.SearcherSingleIndex
 import com.algolia.instantsearch.helper.searcher.addFacet
 import com.algolia.search.model.response.ResponseSearch
 
-
 internal data class HierarchicalConnectionSearcher(
     private val viewModel: HierarchicalViewModel,
     private val searcher: SearcherSingleIndex
@@ -17,9 +16,19 @@ internal data class HierarchicalConnectionSearcher(
             val facets = response.hierarchicalFacetsOrNull
                 ?: response.facetsOrNull?.filter { it.key == viewModel.hierarchicalAttributes.first() } ?: mapOf()
 
-            viewModel.tree.value = viewModel.hierarchicalAttributes.mapNotNull { facets[it] }.flatten().toNodes()
+            viewModel.tree.value = viewModel.hierarchicalAttributes
+                .mapNotNull { facets[it] }
+                .flatten()
+                .toNodes(selectedHierarchicalValue)
         }
     }
+
+    private val selectedHierarchicalValue: String?
+        get() {
+            val value = viewModel.selections.value
+            if (value.isEmpty()) return null
+            return value.joinToString(viewModel.separator)
+        }
 
     init {
         searcher.query.addFacet(*viewModel.hierarchicalAttributes.toTypedArray())