+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.5.0
+
+## Added
+- Support and target Android API 30
+- Add selection capability to the hierarchical tree nodes
+- Filter range dynamic behavior based on a searcher state
+
 # 2.4.0
 
 ## Added